<refactor> Метод updateUser в классе UserInMemoryStorage упрощен, логика вынесена в приватные методы, добавлена валидация email через аннотации <refactor> Упразднен класс counter, взамен в классе UserInMemoryStorage и ItemInMemoryStorage добавлено поле counter и метод increaseCounter <refactor> Внутри папок Item и User классы также распределены по соответствующим папка для упрощения логики хранения. <refactor> Прочие мелкие улучшения
