--- src/main/java/ru/practicum/shareit/Counter.java
@@ -1,12 +1,0 @@
-package ru.practicum.shareit;
-
-import org.springframework.stereotype.Component;
-
-@Component
-public class Counter {
-    private long id = 0;
-
-    public long increaseId() {
-        return ++id;
-    }
-}
