+++ src/main/java/ru/practicum/shareit/item/ItemConstants.java
@@ -1,0 +1,5 @@
+package ru.practicum.shareit.item;
+
+public class ItemConstants {
+    public static final String X_SHADER_USER_ID = "X-Sharer-User-Id";
+}
