+++ src/main/java/ru/practicum/shareit/user/storage/UserStorage.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.user;
+package ru.practicum.shareit.user.storage;
 
 import ru.practicum.shareit.user.model.User;
